requirements.txt
fastapi
uvicorn
pydantic
streamlit
faster-whisper
sentence-transformers
faiss-cpu
requests
python-multipart
pydub
